requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain_text_splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-ollama
langchain-core


